requirements.txt
# Dealer Scraper Requirements

# Browser automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# HTTP requests (for API-based scraping)
requests>=2.31.0

# Zip code database for coverage generation
pgeocode>=0.4.0
pandas>=2.0.0

# Web search for finding dealer locator pages
duckduckgo-search>=6.0.0

# Configuration file support
pyyaml>=6.0
